Drop followed_system_boot and fix unconvert lint failure
The followed_system_boot field inferred whether the agent came up as part of
the OS boot from a 2 minute window, which is wrong on machines that take
longer to start the agent. boot_time and system_uptime already carry the
information needed to tell an agent restart from a device reboot.

Also drops the int64 widening of unix.Sysinfo_t.Uptime, which golangci-lint's
unconvert flagged on linux/amd64; time.Duration conversion accepts the int32
form on 32 bit platforms directly.

+++ utils/boot_linux.go
@@ -16,8 +16,9 @@ func systemBootTime() (time.Time, error) {
 	if err := unix.Sysinfo(&info); err != nil {
 		return time.Time{}, errw.Wrap(err, "getting sysinfo")
 	}
-	// Uptime is int32 on 32 bit platforms, so widen before converting to a Duration.
-	return time.Now().Add(-time.Duration(int64(info.Uptime)) * time.Second), nil
+	// Uptime is int32 on 32 bit platforms and int64 on 64 bit ones; converting to a
+	// Duration handles both.
+	return time.Now().Add(-time.Duration(info.Uptime) * time.Second), nil
 }
 
 func systemIsShuttingDown(ctx context.Context, logger logging.Logger) bool {
